Adding updated car stats logic to car class

+++ Assets/Scripts/Constants/CarLevels/CarStats.cs
@@ -1,15 +1,15 @@
 public readonly struct CarStats
 {
-    public double Speed { get; init; }
-    public double Health { get; init; }
+    public float Speed { get; init; }
+    public float Health { get; init; }
     public int Ammo { get; init; }
-    public double Damage { get; init; }
+    public float Damage { get; init; }
     public int BulletsPerShot { get; init; }
     /** Time between each attack, in seconds */
-    public double AttackRate { get; init; }
-    public double Spread { get; init; }
+    public float AttackRate { get; init; }
+    public float Spread { get; init; }
 
-    public CarStats(double speed, double health, int ammo, double damage, double attackRate, double spread) {
+    public CarStats(float speed, float health, int ammo, float damage, float attackRate, float spread) {
         Speed = speed;
         Health = health;
         Ammo = ammo;
@@ -19,7 +19,7 @@ public readonly struct CarStats
         Spread = spread;
     }
 
-    public CarStats(double speed, double health, int ammo, double damage, int bulletsPerShot, double attackRate, double spread) {
+    public CarStats(float speed, float health, int ammo, float damage, int bulletsPerShot, float attackRate, float spread) {
         Speed = speed;
         Health = health;
         Ammo = ammo;
